backend/requirements.txt
django==4.2.15              # Framework principal
djangorestframework         # API REST pour Django
django-cors-headers         # Gestion des CORS
channels                    # WebSockets avec Django
channels-redis              # Backend Redis pour Channels
daphne                      # Serveur ASGI
psycopg2-binary             # Connecteur PostgreSQL
python-dotenv               # Chargement des variables d'environnement
ccxt                        # API d'échanges crypto
redis>=4.3.4                # Client Redis avec support asyncio
websockets                  # Support WebSocket pur Python
pandas-ta-classic           # Indicateurs techniques (trading)
cryptography                # Chiffrement des clés API (Fernet)